fix: Percent slides width human-readable in config
It looks like QSetting() does not handle float the way I thought, and it
was correctly saving/loading values to/from the configuration file, but
in a non-humain readable and editable way. For example, the value 0.55
(for 55%) was saved as:
currentSlidePrcentWidth=@Variant(\0\0\0\x87?\x19\x99\x9a)

Changing from float to double allows to have something more human-ready:
currentSlidePrcentWidth=0.55

+++ parameters.cpp
@@ -208,12 +208,12 @@ QString Parameters::getPdfFileName()
     return this->pdfFileName;
 }
 
-void Parameters::setCurrentSlidePrcentWidth(float v)
+void Parameters::setCurrentSlidePrcentWidth(double v)
 {
     this->currentSlidePrcentWidth = v;
 }
 
-float Parameters::getCurrentSlidePrcentWidth()
+double Parameters::getCurrentSlidePrcentWidth()
 {
     return this->currentSlidePrcentWidth;
 }
